Começar na GUI
Inserido JavaFx lib no projeto, começar na GUI quando prontos.

+++ src/pt/isec/pa/javalife/model/data/elements/Flora.java
@@ -3,7 +3,7 @@ package pt.isec.pa.javalife.model.data.elements;
 import pt.isec.pa.javalife.model.data.Area;
 
 public sealed class Flora extends ElementoBase implements IElementoComForca, IElementoComImagem permits Erva {
-    private int forca;
+    private double forca;
     private String imagem;
 
     /*public Flora(String imagem) {
